Move to grabbing owner_id and exporting everything over GRPC lib
