+++ schematic/__init__.py
@@ -5,8 +5,12 @@ from typing import Dict, List
 
 import pkg_resources
 from opentelemetry import trace
-from opentelemetry.exporter.otlp.proto.http.trace_exporter import OTLPSpanExporter
+from opentelemetry._logs import set_logger_provider
+from opentelemetry.exporter.otlp.proto.grpc._log_exporter import OTLPLogExporter
+from opentelemetry.exporter.otlp.proto.grpc.trace_exporter import OTLPSpanExporter
 from opentelemetry.instrumentation.flask import FlaskInstrumentor
+from opentelemetry.sdk._logs import LoggerProvider, LoggingHandler
+from opentelemetry.sdk._logs.export import BatchLogRecordProcessor
 from opentelemetry.sdk.resources import SERVICE_NAME, Resource
 from opentelemetry.sdk.trace import TracerProvider
 from opentelemetry.sdk.trace.export import (
@@ -19,8 +23,6 @@ from opentelemetry.sdk.trace.sampling import ALWAYS_OFF
 from synapseclient import Synapse
 from werkzeug import Request
 
-from schematic.configuration.configuration import CONFIG
-from schematic.loader import LOADER
 from schematic_api.api.security_controller import info_from_bearer_auth
 
 Synapse.allow_client_caching(False)
@@ -76,6 +78,28 @@ def set_up_tracing() -> None:
         trace.get_tracer_provider().add_span_processor(processor)
     else:
         trace.set_tracer_provider(TracerProvider(sampler=ALWAYS_OFF))
+
+
+def set_up_logging() -> None:
+    """Set up logging to export to OTLP."""
+    logging_export = os.environ.get("LOGGING_EXPORT_FORMAT", None)
+    logging_service_name = os.environ.get("LOGGING_SERVICE_NAME", "schematic-tests")
+    logging_instance_name = os.environ.get("LOGGING_INSTANCE_NAME", "local")
+    if logging_export == "otlp":
+        resource = Resource.create(
+            {
+                "service.name": logging_service_name,
+                "service.instance.id": logging_instance_name,
+            }
+        )
+
+        logger_provider = LoggerProvider(resource=resource)
+        set_logger_provider(logger_provider=logger_provider)
+
+        exporter = OTLPLogExporter()
+        logger_provider.add_log_record_processor(BatchLogRecordProcessor(exporter))
+        handler = LoggingHandler(level=logging.NOTSET, logger_provider=logger_provider)
+        logging.getLogger().addHandler(handler)
 
 
 def request_hook(span: Span, environ: Dict) -> None:
